Shrubbery.BranchIndex

Description

This module provides types and functions for finding and using the index of a type within a type-level list.

Documentation

data BranchIndex t (types :: [Type])

A BranchIndex is an zero-based index into a list of types for which the type at the index is known. This type is used with Branches to perform efficient branching at runtime.

firstIndexOfType :: (KnownNat branchIndex, branchIndex ~ FirstIndexOf t types) => BranchIndex t types

Finds the first index of a type in a list of types. The type list and the type being searched for must be inferrable based on the usage, or be supplied via TypeApplications.

indexOfTypeAt :: (KnownNat branchIndex, t ~ TypeAtIndex branchIndex types) => proxy branchIndex -> BranchIndex t types

Builds an index based on an known index value from the type system, associating it with the appropriate type from the list. The index is specified via a proxy value.

branchIndexToInt :: BranchIndex t types -> Int

Retrieves the Int representation of the branch index. It it up to the caller to use this integer in a responsible fashion ;)

data TypeZipper (front :: [Type]) (focus :: Type) (back :: [Type])

TypeZipper facilitates interating through a list of known types while keeping track of the index as you go. You might need this if you're keeping track of a list of operations that use shrubbery functions that require a branch index, such as unifyWithIndex.

Note: Types are added to the top of the front list as the are encountered, so the list may be reversed from you expected. See the type of nextZipper.

startZipper :: TypeZipper '[] first back

Initializes a TypeZipper at the start of the type list

moveZipperNext :: TypeZipper front focus (next ': back) -> TypeZipper (focus ': front) next rest

Moves the zipper to the next type in the list. The currently focused item is added to the to the beginning of the front list.

indexOfFocusedType :: TypeZipper front focus back -> BranchIndex t (ZippedTypes front focus back)

Builds a BranchIndex for the currently focused type of the zipper. This index can then be used with other functions such as unifyWithIndex.