Shrubbery.Parser

Description

This module provides facilities for parsing input values into sum types that implement the Unification type classes. A Parser is constructed by giving a handler for each type in the union that attempts to parse the input value into that type. These handlers must be given in the same order as the types in the type-list that describes the sum type.

For instance:

 parser :: Parser Maybe String [Int, Bool, String]
 parser =
     parseOption (readMaybe :: String -> Maybe Int)
   $ parseOption (readMaybe :: String -> Maybe Bool)
   $ parseOption (Just :: Strng -> Maybe String)
   $ parseEnd
 

Or, if you prefer to use TypeApplications,

 parser :: Parser Maybe String [Int, Bool, String]
 parser =
     parseOption @Int readMaybe
   $ parseOption @Bool readMaybe
   $ parseOption @String Just
   $ parseEnd
 

When a Parser is run via parse, it returns the parsing results for all the options, allowing the caller to choose how to handle them. For instance, using the parser above one might write the following code to pick the first successful option from the list:

 case catMaybes (parse parser someInputString) of
   (first:_) ->
     Just first

   [] ->
     Nothing
 

Documentation

data Parser f input (types :: [Type])

A Parser allows an input value to be examined to see whether one or more underlying types can be constructed from it, which are then embedded in a union. The resulting type can be any type that provides an instance of Unification, such as Union.

parse :: (Unification sum, Functor f) => Parser f input (BranchTypes sum) -> input -> [f sum]

Executes a Parser with a given input value, producing the list of options returned be each function given to parseOption. The returned list will always have a number of items equal to the number of options added to the Parser.

The caller must decide how to choose one or more results from the list of options however is desired.

parseOption :: forall a f input rest. (input -> f a) -> Parser f input rest -> Parser f input (a ': rest)

Adds an option to the Parser. Every option in the Parser will be given a chance to examine the input and possibly produce a value of the type it produces. IF the option cannot produce a value its type from the input, it should return whatever the appropriate value for functor f is. For example, if Maybe is used as f, the function would produce Nothing when given input it does not want to handle.

Any values a that are produced by the function will be unified with the other values according the Unification sum type that is expected when parse is used.

parseEnd :: Parser f input '[]

Represents the base case of where no parsing options are present. You can use this as the second argument to parseOption when have no other options to specify.