Shrubbery.TypeList

Description

This module provides type families that are useful for working with type-level lists. Usually you won't need to use these types directly as while using this package, but they may occasionally come in useful if you find yourself writing your own helper functions.

Documentation

type FirstIndexOf t types = FirstIndexOfWithMsg t types (NotAMemberMsg t types)

This type finds the first index of a type in the given list as a type-level natural. It is implemented as a type synonym around FirstIndexOfWithMsg to provide a nice error message when the given type is not found in the list.

type family FirstIndexOfWithMsg t (types :: [Type]) (errMsg :: ErrorMessage) where ...

Finds the first index of a type in the given list as a type-level natural.

The last argument of the type family is the error message that will be given if the type is not found. Normally you will just use the FirstIndexOf synonym, but you can use this instead if you want to provide a customized error message.

Equations

FirstIndexOfWithMsg _ '[] errMsg = TypeError errMsg

FirstIndexOfWithMsg t (t ': _) _ = 0

FirstIndexOfWithMsg t (_ ': rest) errMsg = 1 + FirstIndexOfWithMsg t rest errMsg

type TypeAtIndex n types = TypeAtIndexWithMsg n types (OutOfBoundsMsg n types)

This type finds the type at the given index in a list of types. It is implemented as a type synonym around TypeAtIndexWithMsg to provide a nice error message when the given type is not found in the list.

type family TypeAtIndexWithMsg (n :: Nat) (types :: [Type]) (errMsg :: ErrorMessage) where ...

Finds the type at the given index in a list of types.

The last argument of the type family is the error message that will be given if the type is not found. Normally you will just use the TypeAtIndex synonym, but you can use this instead if you want to provide a customized error message.

Equations

TypeAtIndexWithMsg 0 '[] errMsg = TypeError errMsg

TypeAtIndexWithMsg 0 (t ': _) _ = t

TypeAtIndexWithMsg n (_ ': rest) errMsg = TypeAtIndexWithMsg (n - 1) rest errMsg

class KnownLength (types :: [Type]) where

Similar to KnownNat, this class allows length of a type list that is known at compile time to be retrieved at runtime. This functionality is exported as a class rather than a family both to simplify the type signatures of functions that use this knowledge and for situations where consumers would need to turn on UndecidableInstances to use it.

Methods

lengthOfTypes :: proxy types -> Int

type NotAMemberMsg t (types :: [Type]) = ('ShowType t :<>: 'Text " is not a member of ") :<>: 'ShowType types

This is the default error message used by FirstIndexOf when the type is not found.

type OutOfBoundsMsg (index :: Nat) (types :: [Type]) = (((('Text "Index " :<>: 'ShowType index) :<>: 'Text " is out of bounds for the type list of length ") :<>: 'ShowType (Length types)) :<>: 'Text ": ") :<>: 'ShowType types

This is the default error message used by TypeAtIndex when the index is out of bounds.

type family Length (types :: [Type]) :: Nat where ...

Used by KnownLength to calculate the length of a type-level list.

Equations

Length '[] = 0

Length (_ ': rest) = 1 + Length rest

type family ZippedTypes front focus back :: [Type] where ...

Equations

ZippedTypes '[] focus back = focus ': back

ZippedTypes (a ': rest) focus back = ZippedTypes rest a (focus ': back)